requirements.txt
bert_extractive_summarizer==0.10.1
Flask==3.1.0
folium==0.19.4
geopy==2.4.1
goose3==3.1.19
matplotlib==3.10.0
nltk==3.9.1
numpy==2.2.1
openai==1.59.7
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
spacy==3.8.3
torch==2.3.1
